requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
pillow
python-dotenv
numpy
pyjwt
